COMP: removing compiler warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=6545
git-svn-id: http://svn.slicer.org/Slicer4/trunk@6545 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Volumes/vtkSlicerDWITestingWidget.cxx
@@ -310,7 +310,7 @@ void vtkSlicerDWITestingWidget::ProcessWidgetEvents (vtkObject *caller, unsigned
     {
     //set new resolution
     this->VolumesGUI->GetdtiVDW()->GetGlyphDisplayWidget()->SetGlyphRosolution(
-      this->GlyphResolutionScale->GetWidget()->GetValue());
+        (int)(this->GlyphResolutionScale->GetWidget()->GetValue()));
     }
 
   //view of tracts
@@ -350,7 +350,7 @@ void vtkSlicerDWITestingWidget::CreateGlyphs(vtkKWCheckButton *calledGlyph)
     this->Application->GetApplicationGUI()->GetApplicationLogic()->PropagateVolumeSelection();
     //adjust resolution
     this->VolumesGUI->GetdtiVDW()->GetGlyphDisplayWidget()->SetGlyphRosolution(
-      this->GlyphResolutionScale->GetWidget()->GetValue());
+      (int)(this->GlyphResolutionScale->GetWidget()->GetValue()));
     }
   else
     {
